Renames to Redistribute, places in .transforms
